Firestore

+++ app/src/main/java/app/resta/com/restaurantapp/util/GroupPositionComparator.java
@@ -13,12 +13,12 @@ public class GroupPositionComparator implements Comparator<RestaurantItem> {
     @Override
     public int compare(RestaurantItem item1, RestaurantItem item2) {
 
-        Integer position1 = -1;
+        Long position1 = -1l;
         if (item1 != null) {
             position1 = item1.getPosition();
         }
 
-        Integer position2 = -1;
+        Long position2 = -1l;
         if (item2 != null) {
             position2 = item2.getPosition();
         }
